miercuri 21:32

+++ vma.h
@@ -51,5 +51,5 @@ void free_block(arena_t* arena, const uint64_t address);
 void read(arena_t* arena, uint64_t address, uint64_t size);
 void write(arena_t* arena, const uint64_t address,  const uint64_t size, int8_t *data);
 void pmap(const arena_t* arena);
-int interpretare_string(char string[100]);
+int8_t interpretare_string(char string[100]);
 void mprotect(arena_t* arena, uint64_t address, int8_t *permission);
